+++ src/config/configSchema.ts
@@ -97,6 +97,7 @@ export type ConfigurationSchema = {
   manualRecordUpload: boolean;
   firstTimeSetup: boolean;
   chatUserNameAgreed: string;
+  validateLogPaths: boolean;
 };
 
 export type ConfigurationSchemaKey = keyof ConfigurationSchema;
@@ -606,5 +607,10 @@ export const configSchema = {
     description: Phrase.Unknown, // Not actually exposed.
     type: 'string',
     default: '',
+  },
+  validateLogPaths: {
+    description: Phrase.ValidateLogPathsDescription,
+    type: 'boolean',
+    default: true,
   },
 };