+++ src/localisation/german.ts
@@ -501,6 +501,8 @@ const GERMAN: Translations = {
   [Phrase.SelectOptions]: 'Optionen wählen',
   [Phrase.SelectLanguage]: 'Sprache wählen',
   [Phrase.ChatDeleteMessageTooltip]: 'Diese Chat-Nachricht löschen',
+  [Phrase.ValidateLogPathLabel]: 'Logpfade überprüfen',
+  [Phrase.ValidateLogPathsDescription]: 'Warcraft Recorder überprüft, ob die von Ihnen festgelegten Logpfade auf eine unterstützte WoW-Installation verweisen. Sie können diese Überprüfung deaktivieren, um nicht unterstützte Spielmodi festzulegen. Deaktivieren Sie dies auf eigenes Risiko.',
 };
 
 export default GERMAN;