+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Changed
 ### Added
+- A switch to disable log path validation so I don't need to constantly update the supported flavours and things like classic anniversary realms can work.
 ### Fixed
 
 ## [7.4.0] - 2026-01-08